+++ installation-bundle/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Contao installation bundle change log
 
+### DEV
+
+ * Use the table options instead of the default table options to compare engine and collation.
+
 ### 4.5.0-RC1 (2017-12-12)
 
  * Check the MySQL version and the configured database options.